Revert to Settings class to helper module
