+++ tests/test_helper.py
@@ -5,7 +5,6 @@ import os
 import requests
 from requests.models import Response
 
-from aqt.settings import Settings
 from aqt import helper
 
 
@@ -52,11 +51,11 @@ def test_helper_altlink(monkeypatch):
 
 
 def test_settings(tmp_path):
-    Settings.load_settings(
+    helper.Settings.load_settings(
         os.path.join(os.path.dirname(__file__), "data", "settings.ini")
     )
-    assert Settings.concurrency == 3
-    assert "http://mirror.example.com" in Settings.blacklist
+    assert helper.Settings.concurrency == 3
+    assert "http://mirror.example.com" in helper.Settings.blacklist
 
 
 def mocked_iter_content(chunk_size):
